+++ index.js
@@ -135,8 +135,14 @@ module.exports = function(tilelive, options) {
               return next(err);
             }
 
-            vt.addImage(im.encodeSync("webp"), src.info.id.replace(/\./g, "_"));
-            return next(null, vt, headers);
+            return async.waterfall([
+              async.apply(im.encode, "webp"),
+              function(img, done) {
+                vt.addImage(img, src.info.id.replace(/\./g, "_"));
+
+                return done(null, vt, headers);
+              }
+            ], next);
           });
         }
       ], function(err, tile, headers) {